requirements.txt
alembic==1.6.5
click==8.0.1
Flask==2.0.1
Flask-Admin==1.6.0
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-dotenv==0.18.0
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.20
Werkzeug==2.0.1
WTForms==3.0.1
